refactor(analytics): use hooks and resolvers
