+++ packages/server-core/src/analytics/analytics/analytics.class.ts
@@ -23,9 +23,8 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { Paginated, Params } from '@feathersjs/feathers'
-import type { KnexAdapterOptions } from '@feathersjs/knex'
-import { KnexAdapter } from '@feathersjs/knex'
+import { Params } from '@feathersjs/feathers'
+import { KnexService } from '@feathersjs/knex'
 
 import {
   AnalyticsData,
@@ -34,94 +33,14 @@ import {
   AnalyticsType
 } from '@etherealengine/engine/src/schemas/analytics/analytics.schema'
 
-import { instanceAttendancePath } from '@etherealengine/engine/src/schemas/networking/instance-attendance.schema'
-import { userPath } from '@etherealengine/engine/src/schemas/user/user.schema'
-import { Knex } from 'knex'
-import { Application } from '../../../declarations'
 import { RootParams } from '../../api/root-params'
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface AnalyticsParams extends RootParams<AnalyticsQuery> {}
 
-/**
- * A class for Analytics service
- */
-
-export class AnalyticsService<T = AnalyticsType, ServiceParams extends Params = AnalyticsParams> extends KnexAdapter<
+export class AnalyticsService<T = AnalyticsType, ServiceParams extends Params = AnalyticsParams> extends KnexService<
   AnalyticsType,
   AnalyticsData,
   AnalyticsParams,
   AnalyticsPatch
-> {
-  app: Application
-
-  constructor(options: KnexAdapterOptions, app: Application) {
-    super(options)
-    this.app = app
-  }
-
-  async find(params?: AnalyticsParams) {
-    if (params?.query!.action === 'dailyUsers') {
-      const limit = params.query!.$limit || 30
-      const returned: Paginated<AnalyticsType> = {
-        total: limit,
-        skip: 0,
-        limit,
-        data: []
-      }
-      const currentDate = new Date()
-      for (let i = 0; i < limit; i++) {
-        const knexClient: Knex = this.app.get('knexClient')
-
-        const instanceAttendance = await knexClient
-          .countDistinct('userId AS count')
-          .table(instanceAttendancePath)
-          .where('createdAt', '>', new Date(new Date().setDate(currentDate.getDate() - (i + 1))).toISOString())
-          .andWhere('createdAt', '<=', new Date(new Date().setDate(currentDate.getDate() - i)).toISOString())
-          .first()
-
-        returned.data.push({
-          id: '',
-          count: instanceAttendance.count,
-          type: '',
-          createdAt: new Date(new Date().setDate(currentDate.getDate() - i)).toDateString(),
-          updatedAt: new Date(new Date().setDate(currentDate.getDate() - i)).toDateString()
-        })
-      }
-      return returned
-    } else if (params?.query!.action === 'dailyNewUsers') {
-      const limit = params.query!.$limit || 30
-      const returned: Paginated<AnalyticsType> = {
-        total: limit,
-        skip: 0,
-        limit,
-        data: []
-      }
-      const currentDate = new Date()
-      for (let i = 0; i < limit; i++) {
-        const knexClient: Knex = this.app.get('knexClient')
-        const newUsers = await knexClient
-          .count('id AS count')
-          .table(userPath)
-          .where('createdAt', '>', new Date(new Date().setDate(currentDate.getDate() - (i + 1))).toISOString())
-          .andWhere('createdAt', '<=', new Date(new Date().setDate(currentDate.getDate() - i)).toISOString())
-          .first()
-
-        returned.data.push({
-          id: '',
-          count: newUsers.count,
-          type: '',
-          createdAt: new Date(new Date().setDate(currentDate.getDate() - i)).toDateString(),
-          updatedAt: new Date(new Date().setDate(currentDate.getDate() - i)).toDateString()
-        })
-      }
-      return returned
-    } else {
-      return await super._find(params)
-    }
-  }
-
-  async create(data: AnalyticsData, params?: AnalyticsParams) {
-    return super._create(data, params)
-  }
-}
+> {}
