+++ packages/server-core/src/analytics/analytics/analytics.hooks.ts
@@ -24,15 +24,22 @@ Ethereal Engine. All Rights Reserved.
 */
 
 import { hooks as schemaHooks } from '@feathersjs/schema'
-import { iff, isProvider } from 'feathers-hooks-common'
+import { discardQuery, iff, isProvider } from 'feathers-hooks-common'
 
 import {
+  AnalyticsType,
   analyticsDataValidator,
   analyticsPatchValidator,
   analyticsQueryValidator
 } from '@etherealengine/engine/src/schemas/analytics/analytics.schema'
+import { instanceAttendancePath } from '@etherealengine/engine/src/schemas/networking/instance-attendance.schema'
+import { userPath } from '@etherealengine/engine/src/schemas/user/user.schema'
+import { Paginated } from '@feathersjs/feathers'
+import { Knex } from 'knex'
+import { HookContext } from '../../../declarations'
 import authenticate from '../../hooks/authenticate'
 import verifyScope from '../../hooks/verify-scope'
+import { AnalyticsService } from './analytics.class'
 import {
   analyticsDataResolver,
   analyticsExternalResolver,
@@ -40,6 +47,70 @@ import {
   analyticsQueryResolver,
   analyticsResolver
 } from './analytics.resolvers'
+
+function checkQueryAction(action: 'dailyUsers' | 'dailyNewUsers') {
+  return (context: HookContext<AnalyticsService>) => context.params.query?.action === action
+}
+
+async function addDailyUsers(context: HookContext<AnalyticsService>) {
+  const limit = context.params.query?.$limit || 30
+  const result: Paginated<AnalyticsType> = {
+    total: limit,
+    skip: 0,
+    limit,
+    data: []
+  }
+
+  const currentDate = new Date()
+  for (let day = 0; day < limit; day++) {
+    const knexClient: Knex = context.app.get('knexClient')
+
+    const instanceAttendance = await knexClient
+      .countDistinct('userId AS count')
+      .table(instanceAttendancePath)
+      .where('createdAt', '>', new Date(new Date().setDate(currentDate.getDate() - (day + 1))).toISOString())
+      .andWhere('createdAt', '<=', new Date(new Date().setDate(currentDate.getDate() - day)).toISOString())
+      .first()
+
+    result.data.push({
+      id: '',
+      count: instanceAttendance.count,
+      type: '',
+      createdAt: new Date(new Date().setDate(currentDate.getDate() - day)).toDateString(),
+      updatedAt: new Date(new Date().setDate(currentDate.getDate() - day)).toDateString()
+    })
+  }
+  context.result = result
+}
+
+async function addDailyNewUsers(context: HookContext<AnalyticsService>) {
+  const limit = context.params.query?.$limit || 30
+  const result: Paginated<AnalyticsType> = {
+    total: limit,
+    skip: 0,
+    limit,
+    data: []
+  }
+  const currentDate = new Date()
+  for (let day = 0; day < limit; day++) {
+    const knexClient: Knex = this.app.get('knexClient')
+    const newUsers = await knexClient
+      .count('id AS count')
+      .table(userPath)
+      .where('createdAt', '>', new Date(new Date().setDate(currentDate.getDate() - (day + 1))).toISOString())
+      .andWhere('createdAt', '<=', new Date(new Date().setDate(currentDate.getDate() - day)).toISOString())
+      .first()
+
+    result.data.push({
+      id: '',
+      count: newUsers.count,
+      type: '',
+      createdAt: new Date(new Date().setDate(currentDate.getDate() - day)).toDateString(),
+      updatedAt: new Date(new Date().setDate(currentDate.getDate() - day)).toDateString()
+    })
+  }
+  context.result = result
+}
 
 export default {
   around: {
@@ -53,7 +124,11 @@ export default {
       () => schemaHooks.validateQuery(analyticsQueryValidator),
       schemaHooks.resolveQuery(analyticsQueryResolver)
     ],
-    find: [],
+    find: [
+      iff(checkQueryAction('dailyUsers'), addDailyUsers),
+      iff(checkQueryAction('dailyNewUsers'), addDailyNewUsers),
+      discardQuery('action')
+    ],
     get: [],
     create: [() => schemaHooks.validateData(analyticsDataValidator), schemaHooks.resolveData(analyticsDataResolver)],
     update: [],
