+++ packages/server-core/src/analytics/analytics/analytics.ts
@@ -44,7 +44,7 @@ export default (app: Application): void => {
     multi: true
   }
 
-  app.use(analyticsPath, new AnalyticsService(options, app), {
+  app.use(analyticsPath, new AnalyticsService(options), {
     // A list of all methods this service exposes externally
     methods: analyticsMethods,
     // You can add additional custom events to be sent to clients here
